+++ libsolidity/formal/SMTEncoder.cpp
@@ -1922,18 +1922,24 @@ smtutil::Expression SMTEncoder::compoundAssignment(Assignment const& _assignment
 
 	auto decl = identifierToVariable(_assignment.leftHandSide());
 
+	TypePointer commonType = Type::commonType(
+		_assignment.leftHandSide().annotation().type,
+		_assignment.rightHandSide().annotation().type
+	);
+	solAssert(commonType == _assignment.annotation().type, "");
+
 	if (compoundToBitwise.count(op))
 		return bitwiseOperation(
 			compoundToBitwise.at(op),
-			decl ? currentValue(*decl) : expr(_assignment.leftHandSide()),
-			expr(_assignment.rightHandSide()),
+			decl ? currentValue(*decl) : expr(_assignment.leftHandSide(), _assignment.annotation().type),
+			expr(_assignment.rightHandSide(), _assignment.annotation().type),
 			_assignment.annotation().type
 		);
 
 	auto values = arithmeticOperation(
 		compoundToArithmetic.at(op),
-		decl ? currentValue(*decl) : expr(_assignment.leftHandSide()),
-		expr(_assignment.rightHandSide()),
+		decl ? currentValue(*decl) : expr(_assignment.leftHandSide(), _assignment.annotation().type),
+		expr(_assignment.rightHandSide(), _assignment.annotation().type),
 		_assignment.annotation().type,
 		_assignment
 	);