+++ test/libsolidity/semanticTests/structs/calldata/calldata_struct_with_bytes_to_memory.sol
@@ -3,14 +3,14 @@ pragma abicoder v2;
 contract C {
     struct S {
         uint256 a;
-        uint256[2] b;
+        bytes b;
         uint256 c;
     }
 
     function f(S calldata c)
         external
         pure
-        returns (uint256, uint256, uint256, uint256)
+        returns (uint256, byte, byte, uint256)
     {
         S memory m = c;
         return (m.a, m.b[0], m.b[1], m.c);
@@ -20,4 +20,4 @@ contract C {
 // ====
 // compileViaYul: also
 // ----
-// f((uint256,uint256[2],uint256)): 42, 1, 2, 23 -> 42, 1, 2, 23
+// f((uint256,bytes,uint256)): 0x20, 42, 0x60, 23, 2, "ab" -> 42, "a", "b", 23