+++ test/libsolidity/semanticTests/uninitializedFunctionPointer/invalidStoredInConstructor.sol
@@ -19,5 +19,7 @@ contract Test {
     new C();
   }
 }
+// ====
+// compileViaYul: also
 // ----
 // f() -> FAILURE, hex"4e487b71", 0x51