+++ test/libsolidity/semanticTests/viaYul/array_memory_as_parameter.sol
@@ -18,7 +18,7 @@ contract C {
 	}
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // ----
 // test(uint256,uint256): 0, 0 -> FAILURE, hex"4e487b71", 0x32
 // test(uint256,uint256): 1, 0 -> 1