+++ test/libsolidity/semanticTests/viaYul/array_memory_index_access.sol
@@ -22,7 +22,7 @@ contract C {
 	}
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // ----
 // index(uint256): 0 -> true
 // index(uint256): 10 -> true