+++ test/libsolidity/semanticTests/viaYul/assert.sol
@@ -14,7 +14,7 @@ contract C {
     }
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // compileToEwasm: also
 // ----
 // f(bool): true -> true