+++ test/libsolidity/semanticTests/viaYul/conversion/implicit_cast_local_assignment.sol
@@ -8,7 +8,7 @@ contract C {
     }
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // compileToEwasm: also
 // ----
 // f() -> 0x78