+++ test/libsolidity/semanticTests/viaYul/function_entry_checks.sol
@@ -17,7 +17,7 @@ contract C {
     }
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // compileToEwasm: also
 // ----
 // f() -> 0