+++ test/libsolidity/semanticTests/viaYul/local_address_assignment.sol
@@ -5,7 +5,7 @@ contract C {
     }
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // compileToEwasm: also
 // ----
 // f(address): 0x1234 -> 0x1234