+++ test/libsolidity/semanticTests/viaYul/loops/break.sol
@@ -24,7 +24,7 @@ contract C {
     }
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // compileToEwasm: also
 // ----
 // f() -> 2