+++ test/libsolidity/semanticTests/viaYul/loops/return.sol
@@ -27,7 +27,7 @@ contract C {
     }
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // compileToEwasm: also
 // ----
 // f() -> 1