+++ test/libsolidity/semanticTests/viaYul/smoke_test.sol
@@ -1,7 +1,7 @@
 contract C {
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // compileToEwasm: also
 // allowNonExistingFunctions: true
 // ----