+++ test/libsolidity/semanticTests/viaYul/storage/mappings.sol
@@ -29,7 +29,7 @@ contract C {
     }
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // ----
 // test_simple(uint256): 0 -> 3, 4, 5
 // test_simple(uint256): 1 -> 3, 4, 5