+++ test/libsolidity/semanticTests/viaYul/virtual_functions.sol
@@ -24,7 +24,7 @@ contract C is X {
     }
 }
 // ====
-// compileViaYul: true
+// compileViaYul: also
 // compileToEwasm: also
 // ----
 // f() -> 3